Guitar Pro : the data types for initial key and octave values were incorrect (opposite)
git-svn-id: http://svn.powertabs.org/svn/ptedev/trunk@498 5b1460fe-822e-4aa1-ac1b-8a6412c92c8c

+++ source/formats/guitar_pro/guitarproimporter.cpp
@@ -47,8 +47,8 @@ std::shared_ptr<PowerTabDocument> GuitarProImporter::load(const std::string& fil
 
     readStartTempo(stream, score);
 
-    /*int8_t initialKey = */stream.read<int8_t>();
-    stream.read<uint32_t>(); // octave
+    /*int32_t initialKey = */stream.read<int32_t>();
+    stream.read<uint8_t>(); // octave
 
     const std::vector<Gp::Channel> channels = readChannels(stream);
